add memo & simple fix

+++ 389/A_9x9.cpp
@@ -4,8 +4,10 @@ using namespace std;
 int main() {
   string S;
   cin >> S;
-  int x = (int)S.at(0) - 48;
-  int y = (int)S.at(2) - 48;
+  int x = S.at(0) - 48;
+  // int x = S.at(0) - '0';
+  int y = S.at(2) - 48;
+  // int y = S.at(2) - '0'; // '0'を引くのでもいい
   cout << x * y << endl;
   return 0;
 }
